:recycle:(chore) move folders and files to suit the DDD architecture
