+++ server/src/modules/users/repositories/UsersRepository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import User from '../models/User';
+import User from '../entities/User';
 
 @EntityRepository(User)
 class UsersRepository extends Repository<User> {
