+++ server/src/modules/users/services/UpdateUserAvatarService.ts
@@ -3,10 +3,10 @@ import path from 'path';
 import fs from 'fs';
 import { StatusCodes } from 'http-status-codes';
 import UsersRepository from '../repositories/UsersRepository';
-import uploadConfig from '../config/upload-config';
+import uploadConfig from '../../../config/upload-config';
 
-import AppException from '../exceptions/AppException';
-import BaseService from '../common/base.services';
+import AppException from '../../../shared/exceptions/AppException';
+import BaseService from '../../../shared/services/Base';
 
 interface Request {
   user_id: string;
