+++ server/src/shared/exceptions/AppException.ts
@@ -1,5 +1,5 @@
 import { StatusCodes } from 'http-status-codes';
-import { HttpKeys } from '../@types/http-status-codes';
+import { HttpKeys } from '../../@types/http-status-codes';
 
 class AppException {
   constructor(
