+++ server/src/shared/routes/index.ts
