+++ server/src/shared/routes/v1/index.ts
