+++ server/src/shared/routes/v1/sessions.routes.ts
@@ -1,6 +1,6 @@
 import { classToPlain } from 'class-transformer';
 import { Router } from 'express';
-import AuthenticateUserService from '../../services/AuthenticateUserService';
+import AuthenticateUserService from '../../../modules/appointments/services/AuthenticateUserService';
 
 const routes = Router();
