+++ server/src/shared/services/Base.ts
@@ -1,4 +1,4 @@
-import { TFunction } from '../@types/i18next.overrides';
+import { TFunction } from '../../@types/i18next.overrides';
 
 class BaseService {
   constructor(protected readonly t: TFunction) {}
